chore: fix broken tests
