+++ code/workspaces/web-app/src/components/notebooks/CreateNotebookForm.js
@@ -4,22 +4,9 @@ import React from 'react';
 import { stackTypes } from 'common';
 import { renderTextField, renderTextArea, renderSelectField, renderAdornedTextField, CreateFormControls } from '../common/form/controls';
 import { syncValidate, asyncValidate } from './newNotebookFormValidator';
+import getUrlNameStartEndText from '../../core/urlHelper';
 
 const { ANALYSIS, getStackSelections } = stackTypes;
-
-export function getUrlNameStartEndText(projectKey, windowLocation) {
-  const separator = '.';
-  const restHostname = windowLocation.hostname.split(separator).slice(1);
-  const startText = `${windowLocation.protocol}//${projectKey}-`;
-
-  let endText = `${separator}${restHostname.join(separator)}`;
-
-  if (windowLocation.hostname === 'localhost') {
-    endText = '.datalabs.localhost';
-  }
-
-  return { startText, endText };
-}
 
 const CreateNotebookForm = (props) => {
   const { handleSubmit, cancel, submitting, dataStorageOptions, projectKey } = props;
