+++ code/workspaces/web-app/src/components/notebooks/CreateNotebookForm.spec.js
@@ -1,41 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { PureCreateNotebookForm, getUrlNameStartEndText } from './CreateNotebookForm';
-
-describe('getUrlNameStartEndText', () => {
-  it('returns correct values when on localhost', () => {
-    const windowLocation = {
-      protocol: 'https:',
-      hostname: 'localhost',
-    };
-    const { startText, endText } = getUrlNameStartEndText(windowLocation);
-
-    expect(startText).toEqual('https://testlab-');
-    expect(endText).toEqual('.datalabs.localhost');
-  });
-
-  it('returns correct values when on testlab.datalabs.localhost', () => {
-    const windowLocation = {
-      protocol: 'https:',
-      hostname: 'testlab.datalabs.localhost',
-    };
-    const { startText, endText } = getUrlNameStartEndText(windowLocation);
-
-    expect(startText).toEqual('https://testlab-');
-    expect(endText).toEqual('.datalabs.localhost');
-  });
-
-  it('returns correct values when on datalab.datalabs.nerc.ac.uk', () => {
-    const windowLocation = {
-      protocol: 'https:',
-      hostname: 'datalab.datalabs.nerc.ac.uk',
-    };
-    const { startText, endText } = getUrlNameStartEndText(windowLocation);
-
-    expect(startText).toEqual('https://datalab-');
-    expect(endText).toEqual('.datalabs.nerc.ac.uk');
-  });
-});
+import { PureCreateNotebookForm } from './CreateNotebookForm';
 
 describe('CreateNotebookForm', () => {
   function shallowRender(props) {
@@ -52,6 +17,7 @@ describe('CreateNotebookForm', () => {
       { text: 'First Data Store', value: 'alpha' },
       { text: 'Second Data Store', value: 'beta' },
     ],
+    projectKey: 'project',
   });
 
   beforeEach(() => jest.resetAllMocks());
