+++ code/workspaces/web-app/src/pages/PublishingPage.spec.js
@@ -2,6 +2,8 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import PublishingPage from './PublishingPage';
 
+const match = { params: { projectKey: 'project' } };
+
 it('PublishingPage renders correct snapshot', () => {
-  expect(shallow(<PublishingPage userPermissions={['expectedPermission']} />)).toMatchSnapshot();
+  expect(shallow(<PublishingPage userPermissions={['expectedPermission']} match={match} />)).toMatchSnapshot();
 });
